requirements.txt
anyio==3.3.1
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.15
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.64.0
fonttools==4.33.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
grpcio==1.47.0
h11==0.9.0
httplib2==0.20.4
httptools==0.1.2
idna==2.10
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
# jio-health-toolkit-logging==0.1.16
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.19.1
motor==2.4.0
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.22.3
packaging==21.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-json-logger==2.0.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.3.0
redis==3.5.3
requests==2.23.0
requests-unixsocket==0.2.0
rsa==4.8
semantic-version==2.8.5
Send2Trash==1.8.0
setuptools-rust==0.12.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.15
starlette==0.13.6
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.25.11
uvicorn==0.11.3
uvloop==0.16.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==8.1
